requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit
transformers
torch
torchvision
pillow
requests
timm
